README.md: Fix #28 "Error during vagrant up"
To fix #28, I finally found a good solution that does not require Linux users to install the binary package of the ChefDK. (vagrant-berkshelf doesn't require the entire ChefDK, only berkshelf. Although ChefDK is the preferred way to install berkshelf, only a few Linux distros are supported.)
1st, understand that with current versions of Vagrant, any gems installed by a user or sysadmin with the `gem` or `bundler` commands are ignored. Instead, vagrant installs vagrant-managed gems, and vagrant plugins use vagrant-managed gems. So if users cannot install the binary package of the ChefDK, they need to install berkshelf as a vagrant-managed gem.
The simplest way to install berkshelf as a vagrant-managed gem is to install vagrant-berkshelf 2.0.1 before installing the latest version of vagrant-berkshelf (4.0.1). So that's my solution.